Use SetOptionNative() instead of setting options directly
Setting options directly in buf.Settings without calling SetOption() or
SetOptionNative() is generally not the best idea, since it may not
trigger the needed side effects. In particular, after
https://github.com/zyedidia/micro/pull/3343, directly setting
`tabstospaces` and `tabsize` causes them not being tracked as locally
overridden per buffer, so if we run the `reload` command, it
unexpectedly replaces the autodetected `tabstospaces` and `tabsize`
settings with the default ones.